Fix case that prevents to unflatten some objects
The following example is not possible to unflatten with the previous
version of the library.

{
       "a.1": 1,
       "a.B": 1
}

The code would try to create an array and set value '1' to index 1.
On the next iteration it would try to set the value '1' to index 'x'
which fails, since all indexes should be positive integers.

In this case the element 'a' is of type JSONObject.